requirements.txt
click==7.1.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
SQLAlchemy==1.3.16
Werkzeug==1.0.1
biopython==1.76
PyMySQL==0.9.3